+++ src/functions/mod.rs
@@ -2,3 +2,4 @@ pub mod list;
 pub mod show;
 pub mod position;
 pub mod set_mode;
+pub mod set_scale;